Fixed bugs in noise function and added other noises

+++ drawable.h
@@ -107,7 +107,7 @@ void drawable_update_buffer(Drawable* drawable, uint buffer_id);
 void drawable_draw(Drawable* drawable);
 void drawable_destroy(Drawable* drawable);
 
-void drawable_init(Drawable* drawable, unsigned short* elements, uint elements_number, ArrayBufferDeclaration* declarations, uint declarations_count, Material* material, GLenum mode, Vector3* position, GLuint* textures, uint textures_count, uint flags);
+void drawable_init(Drawable* drawable, unsigned int* elements, uint elements_number, ArrayBufferDeclaration* declarations, uint declarations_count, Material* material, GLenum mode, Vector3* position, GLuint* textures, uint textures_count, uint flags);
 void drawable_rectangle_texture_init(Drawable* drawable, float width, float height, Material* material, GLenum mode, Vector3* position, GLuint* textures, uint textures_count, float* texture_uv, uint flags);
 void drawable_rectangle_init(Drawable* drawable, float width, float height, Material* material, GLenum mode, Vector3* position, uint flags);
